refactor(plugin): split DailyTasksTable.tsx into sub-components (#2458)
Split monolithic 1287 LOC DailyTasksTable.tsx into focused sub-components
under daily-tasks/ directory:

- types.ts (77 LOC) - interfaces and types
- helpers.ts (294 LOC) - pure helper functions
- useDailyTasksSorting.ts (194 LOC) - sorting/filtering hook
- DailyTasksTableHeader.tsx (132 LOC) - header and colgroup rendering
- DailyTasksTableRow.tsx (204 LOC) - row rendering
- DailyTasksTable.tsx (222 LOC) - main component (reduced)
- DailyTasksTableWithToggle.tsx (225 LOC) - toggle wrapper
- index.ts (18 LOC) - barrel re-export

Old file replaced with re-exports for backward compatibility.
All files under 300 LOC. No logic changes.
